refactor: split monolithic types/index.ts into domain-specific files
Split 654-line types/index.ts into 7 domain files:
prisma.ts, scanner.ts, docker.ts, api.ts, context.ts,
domain.ts, legacy.ts. Barrel re-exports preserve all
existing import paths.
